test(render): split expansion coverage by responsibility
